requirements.txt
Django
git+https://github.com/alex/django-taggit.git
git+https://github.com/glemmaPaul/django-taggit-serializer.git
requests
requests-mock
oauth2
coverage==4.4.2
psycopg2
djangorestframework
codecov
youtube_data_api3
ranking_sorting
elasticsearch>=6.0.0,<7.0.0
elasticsearch-dsl
asyncio-nats-client
